+++ Assets/Scripts/Quests/Goals/MoneyCountGoal.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class MoneyCountGoal : QuestGoal
 {
     private DollarsLogic _dollarsLogic;
-    public override int CurrentProgres { get => _dollarsLogic.CurrencyCount; protected set => base.CurrentProgres = value; }
+    public override double CurrentProgres { get => _dollarsLogic.CurrencyCount; protected set => base.CurrentProgres = value; }
     public override QuestGoalType GoalType => QuestGoalType.MoneyCount;
 
     private void Start()
